Added a WholeTextFileSuite, covering more cases from the corresponding RDD version of the option.
